DataPrep_POC_EDA/pyproject.toml
[tool.poetry]
name = "smartlease"
version = "0.1.0"
description = "smartlease - AI powered search system for housing websites"
authors = ["Shubham Agarwal"]
package-mode = false



[tool.poetry.dependencies]
python = ">=3.10,<3.13"
apache-airflow = "^2.7.1"
openai = "0.28"
googlemaps = "^4.10.0"
pinecone-client = "^5.0.1"
pandas = "^2.2.3"
numpy = "^2.2.0"
fastapi = "^0.115.6"
uvicorn = "^0.32.1"
pydantic = "^2.10.3"
streamlit = "^1.41.0"
pymysql = "^1.1.1"
bcrypt = "^4.2.1"
boto3 = "^1.35.80"
requests = "^2.32.3"
python-jose = "^3.3.0"
python-multipart = "^0.0.19"
selenium = "^4.27.1"
pillow = "^11.0.0"
toml = "^0.10.2"
docker = "^6.1.0"
email-validator = "^2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
